Client check for empty copilots list, fixed failure teardown in java test.
Summary: above

Test Plan: unittests

Reviewers: pja, ranji42, dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D33405

+++ src/test/java/org/rocketspeed/IntegrationTest.java
@@ -31,15 +31,19 @@ public class IntegrationTest {
 
   @Before
   public void setUp() throws Exception {
-    testCluster = new LocalTestCluster();
     testFolder = new TemporaryFolder();
     testFolder.create();
+    testCluster = new LocalTestCluster();
   }
 
   @After
   public void tearDown() throws Exception {
-    testFolder.delete();
-    testCluster.close();
+    if (testCluster != null) {
+      testCluster.close();
+    }
+    if (testFolder != null) {
+      testFolder.delete();
+    }
   }
 
   @Test
